refactor(object): Exp, Yeopjeon 이동
