+++ src/main/java/com/balsamic/sejongmalsami/object/postgres/Exp.java
@@ -1,4 +1,4 @@
-package com.balsamic.sejongmalsami.object;
+package com.balsamic.sejongmalsami.object.postgres;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -19,17 +19,16 @@ import lombok.ToString;
 @AllArgsConstructor
 @NoArgsConstructor
 @ToString
-public class Yeopjeon {
+public class Exp {
 
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
   @Column(columnDefinition = "uuid DEFAULT uuid_generate_v4()", updatable = false)
-  private UUID YeopjeonId;
+  private UUID ExpId;
 
   @OneToOne
   private Member member;
 
-  // 총 엽전량
-  private Integer resultYeopjeon;
-
+  // 총 경험치량
+  private Integer resultExp;
 }
